updating installer/log handling
updated how the dpkg check is done to ensure we're not returning
errors during startup. also, changed the download directory name
to installers, since that's a more accurate description.

logging is no longer done to a file, but rather to stdout. this way
the docker daemon log driver can be used directly. stderr is also
directed to stdout so we don't lose it.